feat: now StarSearch is it's own component so it can be embedded
